bugfix: allow the fighter selector chart to update after being reset
Previously, the workaround was to close the settings menu drawer, because reopening it would force the chart to update.
I had thought that the issue was due to the chart being contained within a drawer, but it turns out that the issue was due to some sort of caching behavior for Altair charts when using dash-vega-components.
I couldn't find any info about this caching behavior in the documentation, but I was able to fix it by breaking the cache with a dummy value.